+++ enphase-influx/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### New features
+
+- Added individual microinverter readings under `inverter` measurement.
+
 ## 0.1.0
 
 Initial release.